updated Pluto notebook and adjusted plotter
